Testing final commit

+++ app/webapp.py
@@ -25,56 +25,58 @@ if not os.path.exists(caches_folder):
     os.makedirs(caches_folder)
 
 def session_cache_path():
-    return caches_folder + session.get('uuid')
+    if session.get('uuid'):
+        return caches_folder + session.get('uuid')
+    else:
+        return None
 
 @server_bp.route('/')
 def index():
-    # if not session.get('uuid'):
-    #     # Step 1. Visitor is unknown, give random ID
-    #     session['uuid'] = str(uuid.uuid4())
-    # auth_manager, cache_handler = get_auth_manager(session_cache_path())
+    if not session.get('uuid'):
+        # Step 1. Visitor is unknown, give random ID
+        session['uuid'] = str(uuid.uuid4())
+    auth_manager, cache_handler = get_auth_manager(session_cache_path())
 
-    # if request.args.get("code"):
-    #     # Step 3. Being redirected from Spotify auth page
-    #     auth_manager.get_access_token(request.args.get("code"))
-    #     return redirect('/')
+    if request.args.get("code"):
+        # Step 3. Being redirected from Spotify auth page
+        auth_manager.get_access_token(request.args.get("code"))
+        return redirect('/')
 
-    # if not auth_manager.validate_token(cache_handler.get_cached_token()):
-    #     # Step 2. Display sign in link when no token
-    #     auth_url = auth_manager.get_authorize_url()
-    #     return render_template('signin.html', auth_url = auth_url)
+    if not auth_manager.validate_token(cache_handler.get_cached_token()):
+        # Step 2. Display sign in link when no token
+        auth_url = auth_manager.get_authorize_url()
+        return render_template('signin.html', auth_url = auth_url)
 
-    # # Step 4. Signed in, display data
-    # spotify = spotipy.Spotify(auth_manager=auth_manager)
+    # Step 4. Signed in, display data
+    spotify = spotipy.Spotify(auth_manager=auth_manager)
     
-    # me =spotify.me() 
-    # # me_dict = {'name': me['display_name'], 'img_url': me['images'][0]['url'], 'id': me['id']}
-    # me_dict = {'name': me['display_name'], 'img_url': me['images'][0]['url'], 'id': '11102560850'}
+    me =spotify.me() 
+    # me_dict = {'name': me['display_name'], 'img_url': me['images'][0]['url'], 'id': me['id']}
+    me_dict = {'name': me['display_name'], 'img_url': me['images'][0]['url'], 'id': '11102560850'}
     
-    # json.dump( me_dict, open( "me.json", 'w' ) )
+    json.dump( me_dict, open( "me.json", 'w' ) )
 
-    # # # Creating a daemon to save the users CSV file
-    # if not os.path.exists(f'{csv_folder}/{get_my_id()}/audio_feature_kmean.csv'):
-    #     save_tse_csv = Process( target=get_tsne_csv, args=([spotify]), \
-    #                                                 kwargs={'min_songs_per_playlist':5,'max_songs_per_playlist':100, 'k':10},  daemon=True)
-    #     save_tse_csv.start()
+    # # Creating a daemon to save the users CSV file
+    if not os.path.exists(f'{csv_folder}/{get_my_id()}/audio_feature_kmean.csv'):
+        save_tse_csv = Process( target=get_tsne_csv, args=([spotify]), \
+                                                    kwargs={'min_songs_per_playlist':5,'max_songs_per_playlist':100, 'k':10},  daemon=True)
+        save_tse_csv.start()
 
-    # # #for saved song history csv file
-    # if not os.path.exists(f'{csv_folder}/{get_my_id()}/saved_track_audio_features.csv'):
-    #     save_hist_csv = Process(target=get_saved_track_history_csv, args=([spotify]), \
-    #                                                 kwargs={'ntracks':1000},  daemon=True)
-    #     save_hist_csv.start()
+    # #for saved song history csv file
+    if not os.path.exists(f'{csv_folder}/{get_my_id()}/saved_track_audio_features.csv'):
+        save_hist_csv = Process(target=get_saved_track_history_csv, args=([spotify]), \
+                                                    kwargs={'ntracks':1000},  daemon=True)
+        save_hist_csv.start()
 
-    # if not os.path.exists(f'{csv_folder}/{get_my_id()}/top_5_artists.csv'):
-    #     save_top_artist_csv = Process(target=get_top_artist_csv, args=([spotify]), daemon=True)
-    #     save_top_artist_csv.start()
+    if not os.path.exists(f'{csv_folder}/{get_my_id()}/top_5_artists.csv'):
+        save_top_artist_csv = Process(target=get_top_artist_csv, args=([spotify]), daemon=True)
+        save_top_artist_csv.start()
 
-    # if not os.path.exists(f'{csv_folder}/{get_my_id()}/top_10_tracks.csv'):
-    #     save_top_tracks_csv = Process(target=get_top_tracks_csv, args=([spotify]), daemon=True)
-    #     save_top_tracks_csv.start()
+    if not os.path.exists(f'{csv_folder}/{get_my_id()}/top_10_tracks.csv'):
+        save_top_tracks_csv = Process(target=get_top_tracks_csv, args=([spotify]), daemon=True)
+        save_top_tracks_csv.start()
 
-    # print("\n Done creating CSVs \n")
-    # return render_template('dashboard.html', spotify = spotify, graphJSON=graphJSON)
+    print("\n Done creating CSVs \n")
     return redirect('/dashboard')
 
 
@@ -87,8 +89,9 @@ def dashboard():
 def sign_out():
     try:
         # Remove the CACHE file (.cache-test) so that a new user can authorize.
-        os.remove(session_cache_path())
-        session.clear()
+        if session_cache_path():
+            os.remove(session_cache_path())
+            session.clear()
         # shutil.rmtree('/home/ubuntu/.spotify_caches')
         # -- TODO -- 
     except OSError as e:
@@ -100,62 +103,62 @@ def serve_sphinx_docs():
     return redirect( url_for('static', filename='html/index.html'))
     
 
-@server_bp.route('/update')
-def update():
-    return redirect(url_for('index'))
+# @server_bp.route('/update')
+# def update():
+#     return redirect(url_for('index'))
 
 
-@server_bp.route('/playlists')
-def playlists():
-    auth_manager, cache_handler = get_auth_manager(session_cache_path())
-    if not auth_manager.validate_token(cache_handler.get_cached_token()):
-        return redirect('/')
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-    print(spotify.current_user_playlists() )
-    return spotify.current_user_playlists()
+# @server_bp.route('/playlists')
+# def playlists():
+#     auth_manager, cache_handler = get_auth_manager(session_cache_path())
+#     if not auth_manager.validate_token(cache_handler.get_cached_token()):
+#         return redirect('/')
+#     spotify = spotipy.Spotify(auth_manager=auth_manager)
+#     print(spotify.current_user_playlists() )
+#     return spotify.current_user_playlists()
 
 
-@server_bp.route('/currently_playing')
-def currently_playing():
-    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
-    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
-    if not auth_manager.validate_token(cache_handler.get_cached_token()):
-        return redirect('/')
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-    track = spotify.current_user_playing_track()
-    if not track is None:
-        return track
-    return "No track currently playing."
+# @server_bp.route('/currently_playing')
+# def currently_playing():
+#     cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
+#     auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+#     if not auth_manager.validate_token(cache_handler.get_cached_token()):
+#         return redirect('/')
+#     spotify = spotipy.Spotify(auth_manager=auth_manager)
+#     track = spotify.current_user_playing_track()
+#     if not track is None:
+#         return track
+#     return "No track currently playing."
 
 
-@server_bp.route('/current_user')
-def current_user():
-    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
-    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
-    if not auth_manager.validate_token(cache_handler.get_cached_token()):
-        return redirect('/')
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-    return spotify.current_user()
+# @server_bp.route('/current_user')
+# def current_user():
+#     cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
+#     auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+#     if not auth_manager.validate_token(cache_handler.get_cached_token()):
+#         return redirect('/')
+#     spotify = spotipy.Spotify(auth_manager=auth_manager)
+#     return spotify.current_user()
 
 
-@server_bp.route('/details')
-def details():
-    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
-    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
-    if not auth_manager.validate_token(cache_handler.get_cached_token()):
-        return redirect('/')
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
+# @server_bp.route('/details')
+# def details():
+#     cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
+#     auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+#     if not auth_manager.validate_token(cache_handler.get_cached_token()):
+#         return redirect('/')
+#     spotify = spotipy.Spotify(auth_manager=auth_manager)
 
-    return spotify.devices()
+#     return spotify.devices()
 
-@server_bp.route('/recent')
-def recent():
-    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
-    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
-    if not auth_manager.validate_token(cache_handler.get_cached_token()):
-        return redirect('/')
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
+# @server_bp.route('/recent')
+# def recent():
+#     cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path())
+#     auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+#     if not auth_manager.validate_token(cache_handler.get_cached_token()):
+#         return redirect('/')
+#     spotify = spotipy.Spotify(auth_manager=auth_manager)
 
 
-    # return     spotify.current_user_top_artists( limit=20, offset=0, time_range="long_term")
-    return spotify.current_user_recently_played(limit=50, after=None, before=None)
+#     # return     spotify.current_user_top_artists( limit=20, offset=0, time_range="long_term")
+#     return spotify.current_user_recently_played(limit=50, after=None, before=None)
